style: AppDelegate, SceneDelegate 불필요한 주석제거
