+++ JuiceMaker/JuiceMaker/AppCycle/AppDelegate.swift
@@ -1,0 +1,16 @@
+//
+//  JuiceMaker - AppDelegate.swift
+//  Created by yagom. 
+//  Copyright © yagom academy. All rights reserved.
+// 
+
+import UIKit
+
+@main
+class AppDelegate: UIResponder, UIApplicationDelegate {
+
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        return true
+    }
+}
+
